refactor: always save API key to .netrc after an interactive prompt (#10935)
Changes `auth.prompt_api_key()` to automatically save the input key to .netrc and restructures `_login()` to better accommodate this.

`_login()` was previously saving the value of the `WANDB_API_KEY` environment variable to `.netrc`, which was dubious. It was also not saving a prompted API key to `.netrc` when given `update_api_key=False`, like in `wandb.Api()`.