chore(launch): create file for k8s cleanup (#10901)
Description
-----------
<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->
- Fixes WB-NNNNN
- Fixes #NNNN

This PR creates a file that'll hold the kubernetes cleanup mechanism. I've included this small change so I can use this as the base PR to capture the design.

### Design

The point of this feature is to clean up orphaned resources so they do not block scheduling of new jobs / additional resources.

I wanted the functionality to achieve the following:
1. Additional resources are cleaned up even if the agent restarts
2. Active resources should be safeguarded from being deleted
3. Should be run as a non blocking background task
4. Handles errors well, does not crash agent


### Implementation

**Job resource heirarchy**

This design adds an additional label to resources, `wandb.ai/resource-role` which can either be `primary` or `auxiliary`. This is so we can determine if a job was created as an auxiliary resource or as the run for the job. We need this so we don't mistakenly delete jobs.

**Configurable Cleanup**

This design adds config to manage the k8s resource cleaner:
- `cleanup_initial_delay_seconds`: 60 (default)
  - Env: `WANDB_LAUNCH_CLEANUP_INITIAL_DELAY_SECONDS`
- `cleanup_interval_seconds`: 1200 (20 minutes, default)
  - Env: `WANDB_LAUNCH_CLEANUP_INTERVAL_SECONDS`
- `cleanup_min_resource_age_seconds`: 900 (15 minutes, default)
  - Env: `WANDB_LAUNCH_CLEANUP_MIN_RESOURCE_AGE_SECONDS`
- `WANDB_LAUNCH_MONITORED_NAMESPACES`="default,wandb"  # Default
   - Provides a list of namespaces that will be checked for orphaned resource cleanup
- `WANDB_LAUNCH_CLEANUP_ENABLED`
  - Enables the k8s resource cleaner (disalbed by default)

**Preventing active job resources from being deleted**

We do not want to delete resources that belong to an active job, so we should take into consideration that auxiliary resources are spun up BEFORE the job is.

To prevent mistaken deletion of auxiliary resources, we'll need to use two sources to establish if an auxiliary resource is active or not:
1. The job tracker will have active jobs
2. The k8s resources in the namespace

The deletion logic is as follows:
1. Query k8s for primary jobs using the label `wandb-ai/resource-role=primary`
2. Check the agent's job tracker for jobs active jobs
3. Delete resources if the auxiliary resource id isn't in either source and if the resource is older than 15 minutes (by default, this is configurable though).

**Error handling**

Clean up can fail for various reasons like RBAC for the agent being misconfigured. We should not interrupt the agent because of these failures.

<!--
NEW: We're using a new changelog format that's more useful for users. Please
see CHANGELOG.unreleased.md for details and update on relevant changes such as feature
additions, bug fixes, or removals/deprecations.
-->
- [x] I updated CHANGELOG.unreleased.md, or it's not applicable


Testing
-------
How was this PR tested?

<!--
Ensure PR title compliance with the [conventional commits standards](https://github.com/wandb/wandb/blob/main/CONTRIBUTING.md#conventional-commits)
-->