refactor: define Auth type and core functions (#10944)
Defines the `Auth` type, the `AuthApiKey` implementation, and the core session credential management functions.

Also defines the `AuthIdentityTokenFile` type, although I am not certain about the public interface for it. It's only needed as a bridge between `auth` and `settings` for now; the actual logic around JWTs is still elsewhere.

This PR does not export the new types from `wandb` and does not update old code to use the new functions yet.