refactor: make leveldb.Reader track and return offsets for seeking (#11078)
Adds `NextWithOffset` which returns an offset that can be passed to `SeekRecord`.

This requires two new pieces of state:

- `blockOffset` tracks the absolute offset of the current block
- `nextChunkStart` replaces `j` for determining the start of the next desired chunk. `j` (the end of the current chunk) was used because chunks were always read in order.

I was able to remove the `started` and `processedFirstBlock` state:

- `started` was used by `nextChunk()` to avoid returning EOF on the first call when `r.n == 0` which is the condition that replaces it
- `processedFirstBlock` is not necessary now that `blockOffset` is known; the logic of dealing with the first block is in a new `readBlock()` helper method

Seeking used to work by abusing `j`, but this required calling `nextChunk()` meaning that `SeekRecord()` could return an error when seeking past EOF.

```
// Pseudo-code of original logic.
i, j, n = 0, 0, 0
readChunk() // could error near the end of a file
i, j = c, c
```

This is different from `Seek()` which normally allows seeking past EOF without a problem. The reason this is problematic is because for live files, `readChunk` may return an error other than EOF, complicating the logic for using `SeekRecord()`.

I had to factor out the core logic of `nextChunk` because it failed the cyclomatic complexity lint. I think the result is a lot more readable!

This PR also fixes a subtle bug in the reader, which can only happen when live-syncing a run: `singleReader.Read` could return EOF in the middle of an incomplete record if a chunk ended exactly at a block boundary. The proper error is ErrUnexpectedEOF; otherwise, `io.ReadAll` thinks it successfully read a complete record.

## Testing

The `record_internal_test.go` test suite is very comprehensive and helped catch problems during development. In addition to that:

- Tested with a run that triggers flow control, which seeks
- Tested with `wandb beta sync --live` implemented in a future PR