fix: print a valid URL at the end of a run (#11113)
Print a valid URL at the end of a run instead of a line like

```
wandb: 🚀 View run splendid-fog-121 at:
```

with no URL included.

This message comes from wandb-core which was reading the `run_url` setting. Its value is computed in Python, but only when the entity and project are known. These are available after the first UpsertBucket request when initializing a run, so this was relying on wandb-core and the Python client mutating and passing Settings back and forth, which is why it eventually broke.

Refactored it to not be like that. The `run_url` setting is still problematic; I'll see about removing it. In general, settings should be computed exclusively from inputs.