refactor: use transactionlog in leet (#11103)
Replaces `livereader` / `livestore` by `transactionlog`.

I added a `sync.Pool` to `transactionlog.Reader` that does the same thing as in `LiveStore`. I did some rough benchmarking and found that it improves reading speed quite a lot. I originally didn't want to add it because it's an odd pattern: storing byte buffers in `sync.Pool` is discouraged as it can be a memory leak, but it's valid under the mild assumption that large records in a transaction log are soon followed by more large records (the biggest record is History, and users tend to log the same metrics in each step, so each History record should be a similar size).

I was surprised to find that `LiveStore` is used concurrently. Go tests detected race conditions without the mutex; I think these race conditions must have been present but undetected before this PR.