feat: wandb beta sync --live (#11079)
Fixes WB-27693.

Adds the `--live` option to `wandb beta sync` to allow syncing a run while it is being logged.

This only uploads data as it appears in the run's `.wandb` file, which has some consequences:

- There will be some delay due to polling
- The .wandb file is flushed once every 32 KiB, so the sync may lag behind by an arbitrary amount of time for runs that don't log a lot of data
- Nothing is uploaded at all until the first 32 KiB is written to disk

This behavior can be improved by flushing more frequently.

For now, this should be considered an interactive mode: it may hang if the run stops updating its transaction log before it is done, requiring manually killing the process with Ctrl+C. This can be improved by having all runs create and update a heartbeat file to allow crash detection.