chore: maybe fix debouncedwriter_test flake (#11126)
The `time.Sleep(time.Second)` in the test seems to have resulted in a race with the 1-second rate limit in this [pre-commit run](https://github.com/wandb/wandb/runs/59239895914). Having the test sleep for strictly more than a second should avoid the race.

NOTE: The test doesn't actually sleep as it runs using `testing/synctest`.